# Add noescape and nocallback directive to EVP_PKEY_derive