@doujiang24 2023-11-04
https://go-review.googlesource.com/c/go/+/539235
fyi: it's disabled in go 1.22, will enable in go 1.23.